requirements.txt
Flask==2.3.2
openai
python-dotenv
flask_cors